:truck: Extract component types into separate files
