+++ src/components/HdAlertTypes.js
@@ -1,0 +1,21 @@
+import {
+  checkCircle as successIcon,
+  info as infoIcon,
+  warning as warningIcon,
+  error as errorIcon,
+} from 'homeday-assets';
+
+export default {
+  info: {
+    icon: infoIcon,
+  },
+  success: {
+    icon: successIcon,
+  },
+  warning: {
+    icon: warningIcon,
+  },
+  error: {
+    icon: errorIcon,
+  },
+};
