+++ src/components/HdBadgeTypes.js
@@ -1,0 +1,7 @@
+export default {
+  DEFAULT: '',
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TERTIARY: 'tertiary',
+  DISABLED: 'disabled',
+};
