+++ src/components/HdLinkTypes.js
@@ -1,0 +1,4 @@
+export default [
+  'primary',
+  'secondary',
+];
