+++ src/components/buttons/HdButtonTypes.js
@@ -1,0 +1,7 @@
+export default {
+  PRIMARY: 'primary',
+  SECONDARY: 'secondary',
+  TERTIARY: 'tertiary',
+  FLAT: 'flat',
+  GHOST: 'ghost',
+};
