+++ src/components/notifications/HdNotificationsTypes.js
@@ -1,0 +1,6 @@
+export default {
+  ERROR: 'error',
+  NOTIFICATION: 'notification',
+  INFO: 'info',
+  SUCCESS: 'success',
+};
