+++ src/stories/HdAlert.stories.js
@@ -1,4 +1,5 @@
-import HdAlert, { TYPES } from 'homeday-blocks/src/components/HdAlert.vue';
+import HdAlert from 'homeday-blocks/src/components/HdAlert.vue';
+import TYPES from 'homeday-blocks/src/components/HdAlertTypes';
 import DeprecatedWrapper from 'homeday-blocks/src/storiesWrappers/DeprecatedWrapper';
 
 export default {
