+++ src/stories/HdButton.stories.js
@@ -1,4 +1,5 @@
-import HdButton, { TYPES } from 'homeday-blocks/src/components/buttons/HdButton.vue';
+import HdButton from 'homeday-blocks/src/components/buttons/HdButton.vue';
+import TYPES from 'homeday-blocks/src/components/buttons/HdButtonTypes';
 import { plus as plusIcon } from 'homeday-assets';
 
 export default {
