+++ src/stories/HdLink.stories.js
@@ -1,5 +1,6 @@
 import './styles/StoryContainers.css';
-import HdLink, { TYPES } from 'homeday-blocks/src/components/HdLink.vue';
+import HdLink from 'homeday-blocks/src/components/HdLink.vue';
+import TYPES from 'homeday-blocks/src/components/HdLinkTypes';
 import HdLinkNote from '../notes/HdLink.md';
 
 export default {
