+++ src/stories/HdNotifications.stories.js
@@ -1,6 +1,7 @@
 import HdButton from 'homeday-blocks/src/components/buttons/HdButton.vue';
 import HdNotifications from 'homeday-blocks/src/components/notifications/HdNotifications.vue';
-import HdNotification, { TYPES } from 'homeday-blocks/src/components/notifications/HdNotification.vue';
+import HdNotification from 'homeday-blocks/src/components/notifications/HdNotification.vue';
+import TYPES from 'homeday-blocks/src/components/notifications/HdNotificationsTypes';
 
 export default {
   title: 'Components/Notifications/HdNotifications',
